fix cases of non-modified components
